# Conflicts:
#	src/main/java/com/aston/componentsystem/service/ComponentService.java
#	src/main/resources/application.properties

+++ src/main/java/com/aston/componentsystem/service/ComponentService.java
@@ -2,32 +2,41 @@ package com.aston.componentsystem.service;
 
 import com.aston.componentsystem.model.Component;
 import com.aston.componentsystem.repository.ComponentRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
-@RequiredArgsConstructor
 public class ComponentService {
+    private final ComponentRepository componentRepository;
 
-    public final ComponentRepository componentRepository;
-
+    public ComponentService(ComponentRepository componentRepository) {
+        this.componentRepository = componentRepository;
+    }
 
     public List<Component> getAllComponents() {
         return componentRepository.findAll();
     }
 
-    public Optional<Component> getComponentById(int id) {
-        return componentRepository.findById(id);
+    public Component getComponentById(int id) {
+        Optional<Component> componentOptional = componentRepository.findById(id);
+        if (!componentOptional.isPresent()) {
+            throw new NullPointerException();
+        }
+        return componentOptional.get();
     }
 
-    public void saveComponent(Component component){
+    public void saveComponent(Component component) {
         componentRepository.save(component);
     }
 
-    public void deleteComponent(int id){
-       componentRepository.deleteById(id);
+    public void deleteComponent(int id) {
+        Optional<Component> component = componentRepository.findById(id);
+        if (!component.isPresent()) {
+            throw new NullPointerException();
+        }
+        componentRepository.deleteById(id);
     }
 }
